+++ src/fmr/perf.gtst.cpp
@@ -10,7 +10,7 @@ TEST(FmrPerf, PerfIntSizes) {
   EXPECT_EQ(sizeof (fmr::perf::Timepoint), sizeof (fmr::perf::Count));
 }
 TEST(FmrPerf, MaxPerfCountGT10G) {
-  EXPECT_GT(std::numeric_limits<fmr::perf::Count>::max(), 10000000000);
+  EXPECT_GT(std::numeric_limits<fmr::perf::Count>::max(), size_t(10000000000));
 }
 int main (int argc, char** argv) {
   return fmr::detail::test:: early_main (&argc, argv);
